Change ramp push threshold to whole degrees, default 1°C
The update threshold was a fractional double defaulting to 0.3°C; changed
to an int (default 1) so it only ever holds whole-degree values.

+++ Vulcano-Control/ViewModels/SettingsViewModel.cs
@@ -16,7 +16,7 @@ public partial class SettingsViewModel : ObservableObject
     private int historyRetentionMinutes;
 
     [ObservableProperty]
-    private double rampPushThresholdCelsius;
+    private int rampPushThresholdCelsius;
 
     [ObservableProperty]
     private int rampMaxPushIntervalSeconds;
